Simplify future type
Clippy complains this type is too complex. This type needs to be this
complex, though. But this is a pretty common combination of types. So
the futures crate has BoxFuture, a type specifically to encapsulate the
pattern of a pinned, boxed future trait object. Using it increases code
clarity, in my opinion.