Un-box HttpRouteHandler::new
I ran clippy, and it gave me
https://rust-lang.github.io/rust-clippy/master/#new_ret_no_self

At first I thought about changing the name of the function, but then I
realized that this change is better. This commit increases the options
for callers. There's no reason to have new return a specific trait
object, when a caller that wants this trait object can get one
themselves.