Improve clarity and efficiency
The body of this macro generates a tuple of futures, and awaits them.
This was previously done manually by expanding code to look like this:

(t1.await?, t2.await?, t3.await?)

This implicitly depends on the tuple's evaluation order, and therefore,
also means that things don't happen in parallel, because we await each
one sequentially.

The futures crate provides two macros for this case: join and try_join.
They take a tuple of futures, but awaits on them in parallel. This makes
the code look a bit more clear, removes the evaluation order dependency,
and possibly improves some performance.