The type checking for parameters could be much more stringent and more (#123)
informative. We'd ideally like to raise issues at compile time, but we
don't actually have the data (e.g. in proc macro context) we'd need to
identify type errors. We definitely do not want to fail after the server
has been running when a particular combination of parameters fails to
deserialize into the given structs. As a middle ground, we fail as the
ApiDescription is built up, before the server is servicing requests.