Remove the identifier restriction on path components (#271)
Dropshot deserializes pairs of path component variable names and actual
path component values into a type parameter to a Path Extractor.
Typically these will be valid Rust identifiers, but they don't need to
be e.g. if the consumer were to use #[serde(rename = "while")] to rename
the field during deserialization. The only restriction we need to
enforce is that the path components have *a* name (i.e. it is
non-empty). A mismatch would be detected at runtime... as happens
already independent of whether renamed fields are valid Rust
identifiers.