Remove max log level features on slog dependency
With these features in place, it wasn't possible for a consumer of
`dropshot` to build a release binary that includes `trace!()` logs.
Cargo features are additive, meaning `release_max_level_debug` was set
for any crate that had dropshot in its dependency tree, and slog's
evaluation of this feature set stops at the highest max level limit.