Add `#[allow(clippy::unused_async)]` to emitted endpoints
We require endpoints to be async, but it's not uncommon to have
endpoints that themselves do not need to `.await` anything. If a user
wants to enable the (off-by-default) `clippy::unused_async` lint, they
would need to add this annotation by hand to each such endpoint, or we
could just do it for them.