[test-utils] add log_prefix_for_test + store name of test in LogContext (#878)
While trying to diagnose https://github.com/oxidecomputer/omicron/issues/4779,
which only occurs in CI, I'm trying to collect ClickHouse logs. To do that we
need a way to uniquely identify the ClickHouse logs per test.

It is possible to derive the log prefix from `log_file_for_test`, but it seems
cleaner to do it this way.