[dropshot] add bad_channel fixtures (#1007)
I'm going to be touching a bunch of this code soon, and I realized we didn't
really have any fixtures to check that we were producing good error messages
for channels. So add some, mirroring the `bad_endpoint` fixtures.

The current bad_channel tests provide a baseline -- we can definitely do better
for some of these tests. For now, we can ensure that we aren't regressing them.