[dropshot] add API traits with support for endpoints (#984)
Add support for a new way to define Dropshot APIs: via a trait. This builds on work done in the prior PRs in this series, such that it reuses as much shared logic from endpoints as possible.

This change is almost entirely added code. It's also mostly tests, with around 2000 lines of "real" code.

See RFD 479 for more information: https://rfd.shared.oxide.computer/rfd/0479