add `recommended_status_code()` to dropshot errors
As I proposed in [this comment][1]. This is intended to be used by user
code that constructs its own error type from Dropshot's errors, to make
it easier to get the same status codes that are used by `HttpError` when
the custom user error type just structures the response differently, or
only wishes to override a small subset of Dropshot errors. Perhaps this
should be a trait eventually --- I'm kinda on the fence about this.

We may also want to do a similar "recommended headers" thing, since some
error conditions are supposed to return specific headers, like which
methods are allowed for a Method Not Allowed error code, or the desired
protocol to upgrade to for a 426 Upgrade Required error.

While I was here, I also changed some error variants to return more
correct status codes --- a bunch of stuff currently just returns 400 Bad
Request when it should really return a more specific error like 426
Upgrade Required etc.

[1]: https://github.com/oxidecomputer/dropshot/pull/1164#discussion_r1830220270