router: Reduce string copies in path traversal
Presently, the router's `input_path_to_segments` function segments the
whole input path, copying each each segment from the input path into a
new owned `String`, and returns a `Vec<String>`. This `Vec` is then
immediately iterated over by reference, and each `&String` segment is
then `to_string()`ed into a _second_ owned `String` that's used when
returning the list of path variables. Path segments which are not
returned (i.e., literals) are still copied into a new `String` which is
used only to look up that segment in the current node's map of edges,
which doesn't require an owned `String` to perform. This all feels a bit
unfortunate, as we are allocating path segments a bunch of times more
than we need to. Ideally, we would iterate over slices borrowed from the
`InputPath`, and allocate `String`s only when necessary in order to
return a path variable's value in the lookup result, or when we
percent-decode something.

This commit makes that change. Now, `input_path_to_segments` returns an
`Iterator<Item = Result<Cow<'_, str>, InputPathError>>`, which borrows
the segment from the path unless it was necessary to allocate a string
to percent-decode the segment. This does mean that we return an
`Iterator` of `Result`s rather than a `Result` and thus have to handle a
potential error every time we get the next segment, but I don't think
that's terrible --- it's kind of the inherent cost of doing path
segmentation lazily. This also means that if we do encounter an error,
or if the path is determined not to route to an endpoint, we don't
continue wasting time segmenting the rest of the path; in the present
implementation, we always segment the whole path up front before
traversing it. I haven't done any real performance analysis of this, but
I imagine the lazy approach is more efficient; it certainly *feels*
nicer. It's also less vulnerable to issues where we spend a bunch of
time segmenting a really long path that will never route to anything.