Allow handlers to return user-defined error types (#1180)
Currently, all endpoint handler functions must return a
`Result<T, HttpError>`, where `T` implements `HttpResponse`. This is
unfortunate, as it limits how error return values are represented in
the API. There isn't presently a mechanism for an endpoint to return
a structured error value of its own which is part of the OpenAPI spec
for that endpoint. This is discussed at length in issue #39.

This branch relaxes this requirement, and instead allows endpoint
handler functions to return `Result<T, E>`, where `E` is any type that
implements a new `HttpResponseError` trait. 

The `HttpResponseError` trait defines how to produce an error response
for an error value. This is implemented by `dropshot`'s `HttpError`
type, but it may also be implemented by user errors. Types implementing
this trait must implement `HttpResponseContent`, to determine how to
generate the response body and define its schema, and they must also
implement a method `HttpResponseError::status_code` to provide the
status code to use for the error response. This is somewhat different
from the existing `HttpCodedResponse` trait, which allows successful
responses to indicate at compile time that they will always have a
particular status code, such as 201 Created. Errors are a bit different:
we would like to be able to return any number of different error status
codes, but would still like to ensure that they represent *errors*, in
order to correctly generate an OpenAPI document where the error schemas
are returned only for error responses (see [this comment][1] for
details). As discussed [here][2], we ensure this by providing new
`ErrorStatusCode` and `ClientErrorStatusCode` types, which are newtypes
around `http::StatusCode` that only contain a 4xx or 5xx status (in the
case of `ErrorStatusCode`), or only contain a 4xx (in the case of
`ClientErrorStatusCode`). These types may be fallibly converted from an
`http::StatusCode` at runtime, but we also provide constants for
well-known 4xx and 5xx statuses, which can be used infallibly. The
`HttpResponseError::status_code` method returns an `ErrorStatusCode`
rather than a `http::StatusCode`, allowing us to ensure that error types
always have error statuses and generate a correct OpenAPI document.

Additionally, while adding `ErrorStatusCode`s, I've gone ahead and
changed the `dropshot::HttpError` type to also use it, and changed the
`HttpError::for_client_error` and `HttpError::for_status` constructors
to take a `ClientErrorStatusCode`. Although this is a breaking change,
it resolves a long-standing issue with these APIs: currently, they
assert that the provided status code is a 4xx internally, which is often
surprising to the user. Thus, this PR fixes #693.

Fixes #39
Fixes #693
Fixes #801

This branch is a second attempt at the change originally proposed in PR
#1164, so this closes #1164. This design is substantially simpler, and
*only* addresses the ability for handler functions to return
user-defined types. Other changes made in #1164, such as a way to
specify a global handler for dropshot-generated errors, and adding
headers to `HttpError` responses, can be addressed separately. For now,
all extractors and internal errors still produce `dropshot::HttpError`s.
A subsequent change will implement a mechanism for providing alternate
presentation for such errors (such as an HTML 404 page).

[1]:
    https://github.com/oxidecomputer/dropshot/issues/39#issuecomment-2483898918
[2]:
    https://github.com/oxidecomputer/dropshot/issues/39#issuecomment-2486602655