fix a few docs bugs and export HttpResponseContent
HttpResponseError is made public at the crate root, but it requires
implementers also implement HttpResponseContent. HttpResponseContent is
*not* public, so implementations of HttpResponseError are constrained to
only being types that are covered by blanket impls of
HttpResponseContent.

i personally don't need to do that, but it makes the docs funky and is
probably an error. it's morally similar to a function returning a type
which is not public.